From a posting from Joe DiMartino dated 4/12/2002: Hello OCF list,
Enclosed, please find the initial draft proposal for the Cluster Event
Notification API (v0.1).

As suggested by Bruce Walker on a previous OCF posting, we have decided
to give only one part of this document at a time for general review.

This document will EVENTUALLY cover:

	a. event notification service
	b. connectivity events and semantics
	c. membership events and semantics
	d. group messaging events and semantics

Currently, this text covers only (a).  The Data Structures section
hints at some events for (b) (c) and (d), but please don't dwell on
that now.

The examples section is intended to increase the understanding of
the event model.  However, there are no examples of in-kernel usage
and three user-level examples are combined into one.

Happy Reading!

Respectfully,
- Ram Pai and Joe DiMartino