Convert Resource Agent API to markdown
Removes the docbook and text format documents describing the
Resource Agent API, and replaces them with a Markdown document.

The benefits of using Markdown over the two previous formats is
that a single source document covers all use cases, making it
easier for people to submit changes to the API, and avoids
potential differences between the document formats.

There are other document markup formats, but none are as widely used
as Markdown.