Add optional attribute 'replaced-by' for parameters/parameter
This attribute allows us to mark a parameter as obsolete one and points us to the replacement. It is required for renaming (= standardizing) names of parameters, so additional tools (e.g. pcs) can properly ignore those parameters if required.

It might be possible that there parameter is completely discarded but that did not happend yet. Currently, similar attribute is used in the fence agents only.

    This attribute is already heavily used so it should be in standard. It
    looks very useful but often it would be preferable to have an option
    to enter more complex queries (e.g enter X or Y)