Add optional attribute 'replaced-by' for parameters/parameter
This attribute allows us to mark a parameter as obsolete one and points us to the replacement. It is required for renaming (= standardizing) names of parameters, so additional tools (e.g. pcs) can properly ignore those parameters if required.

It might be possible that a parameter is completely discarded but that did not happend yet. Currently, similar attribute is used in the fence agents only.