ra/next: Update "deprecated" element to take "replaced-with" as child elements
Previously, the proposed "deprecated" element in resource agent meta-data
took an optional "replaced-with" attribute to indicate another parameter that
should be used instead of the deprecated one. After discussion at
https://github.com/ClusterLabs/OCF-spec/pull/21 , it was decided that having
"replaced-with" as a child element would be preferable, so that multiple
replacements could be specified (for example, separating a deprecated IP:PORT
parameter into separate IP and port number parameters).