ra/next: Deprecate "unique" attribute in favor of new "unique-group"
In the 1.0 version of the standard, each instance of a particular resource type
must have a unique combination of parameter values marked "unique". However,
the Pacemaker cluster resource manager overloaded this attribute to mean
"reloadable", which is an incompatible usage.

This addresses that situation by deprecating "unique" entirely, and instead
adding a new "unique-group" parameter attribute that serves the same function
but allows for multiple groups that each must be unique.

The new attribute is not as versatile as, for example, an RNG-like syntax,
but it is backward-compatible, and a complex syntax would confuse the average
resource agent developer or system administrator, and impose a burden on
meta-data processors.