ra/next: Add "reload-params" action and "reloadable" parameter attribute
Previously, the Pacemaker cluster resource manager misused the (now deprecated)
"unique" parameter attribute to indicate that changes in the parameter value
could take effect via a "reload" action rather than a full stop and start.

Not only was this inconsistent with the standard, but also "reload" was widely
implemented in practice as a way to reload the service externally (for example,
after changes to the service's configuration file), rather than a way to reload
the resource parameters.

This addresses the issue by reserving the "reload" action for its common usage,
and adding a new "reload-params" action to reload all parameters marked with a
new "reloadable" attribute.

An alternative approach would be to require the "reload" action to always
handle both purposes, but this would be inefficient because the two situations
would always be independent (i.e., after editing an external configuration file
in the one case, or after modifying the resource parameters in the other case).