ra/next: Remove editor and update copyright
The resource agent standard is now "edited" via Github pull requests, and is
not assigned to one person in particular, so remove the "Editor" designation.

Similarly, pull requests may be submitted by many authors, so change the
copyright notice to "the Open Cluster Framework project contributors" to
reflect this. Note that copyright notices are intended to indicate, but do not
affect, copyright ownership, which is determined by appropriate laws and
regulations.

Because ClusterLabs and the Open Cluster Framework are informal communities
rather than formally organized entities, it is difficult to come up with
anything more specific.