ra/next: Expand OCF_CHECK_LEVEL to validate-all actions
Problems have occasionally arisen due to the desire to use the validate-all
action in two incompatible contexts: on a cluster node, to determine the
suitability of the node for running the agent; and on any arbitrary host, to
check only the syntax of provided parameters, before applying a new
configuration.

This solves the issue by allowing callers to specify OCF_CHECK_LEVEL with
validate-all actions.