drop references to opencf.org, we lost ownership
Note that the gh-pages branch in this repository is rendered
via https://standards.clusterlabs.org

We may look into updating it.

On that gh-pages branch, there is just one commit from 7 years ago, and mine
now fixing the reference from github.com/OpenCF to
ClusterLabs/OCF-spec/tree/gh-pages, as that seems to be what was actually used.